system_run 分離
