+++ philo/srcs/main.c
@@ -6,35 +6,11 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/10 15:47:33 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/06/17 22:37:21 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/06/17 23:19:18 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "main.h"
-
-void	system_run(t_system *sys)
-{
-	t_ft_msg	*sys_msg;
-
-	ft_actor_start(sys->sv->base);
-	while (true)
-	{
-		sys_msg = sys->notify_inbox->deq(sys->notify_inbox);
-		if (sys_msg && sys_msg->type == ACTOR_START_DONE)
-		{
-			msg_del(&sys_msg);
-			break ;
-		}
-		msg_del(&sys_msg);
-	}
-	while (true)
-	{
-		sys->sv->base->tell(sys->sv->base, msg_new(UPDATE, NULL, NULL));
-		sys_msg = sys->notify_inbox->deq(sys->notify_inbox);
-		if (sys_msg && sys_msg->type == ACTOR_STOP_DONE)
-			break ;
-	}
-}
 
 int	main(int argc, char **argv)
 {
